Разделяет проект на первоначальные модули
